+++ src/main/java/com/spring/boot/application/repository/OpinionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OpinionRepository extends CrudRepository<Opinion,Long> {
+public interface OpinionRepository extends CrudRepository<Opinion,Integer> {
 }
