+++ src/main/java/com/spring/boot/application/service/MusicStoreService.java
@@ -9,13 +9,15 @@ import java.util.Optional;
 
 @Service
 public class MusicStoreService {
+
     @Autowired
     MusicStoreRepository musicStoreRepository;
 
     public Iterable<MusicStore> findAll(){return musicStoreRepository.findAll();}
     public MusicStore save(MusicStore musicStore){return  musicStoreRepository.save(musicStore);}
-    public Optional<MusicStore> findById(Long id){return musicStoreRepository.findById(id);}
+    public Optional<MusicStore> findById(Integer id){return musicStoreRepository.findById(id);}
     public void delete(MusicStore musicStore){musicStoreRepository.delete(musicStore);}
+
 
     //public MusicStore getMusicStore(Long storeId) {
       //  return musicStoreRepository.getMusicStoreBy(storeId);
